Fixed and finalized Mr. Shipper
Finally made Mr. Shipper work the way it was intended. All XMLs and *.cs
files generated are now correct. The assembly version is at 0.9.5.0,
meaning it is 95% done (some of the generated *.cs files have duplicate
entries that must be removed manually).
Renamed TSO MeshViewer to TSO DressUp, and tested Mr. Shipper with
DressUp and the client.