Work on CityRenderer
Code still kind of a mess. Changes:

- Fixed all code - most of the errors in Afr0's attempted port were
accidental int divisions, though one of them was trying to draw a
triangles array as a triangle strip....
- Implemented roads - I didn't actually complete this for the WebGL
version.
- Added functionality that was in the WebGL version that wasn't ported
yet.

I still need to add sanity checks for if the city image uses any unusual
colours or doesn't exist so that people can't crash the game...