Resolved two independent issues
Fixed an language problem where the game would crash if the Norwegian
files aren't installed. This is actually Rhy's fix,  but it was only
applied to his MonoGame port. So, I copied it over.

I also fixed a compiler error where it couldn't locate the Project
Dollhouse font, originally located in outside of the repo.