For the contributors
README.md now gives a slew of new information on how to contribute
project. It also includes a link to a new "up for grabs" tag within
issues to give contributors looking for something to do ... something to
do. Also minor update to Doxyfile that builds outside of Github folder.