Novel: GluonHostPool is a utility to help servers talk to each other. The intention is to replace CityConnections and the admin => shutdown / announce signals with this. It is similar to the CityConnections but it connects to all server types not just city servers.
The process connects to all other servers and monitors those connections. It uses the database hosts table for service discovery but nodes also keep a pool hash which they share. When the pool hash is different it triggers servers to refresh their view of the pool.

Since this approach connects every process to every process its ok for small scale but won't work for lots of servers (erm, maybe up to 10/20 is ok?).