New EOD: Band
To give you the maximum time needed to review and commit this code, I'll
spare you the details!

Suffice to say there are some UI changes I'd like to make to clean it
up.  But it's funcational and ready for test.  I hope I've made enough
checks against race issues.