Roulette Penultimate Patch (#79)
* Roulette EOD

These are all the changes required for roulette to work except the .piff
file for the object.

Users are likely to complain about new players pushing off existing
players when attemption to join. The ability for the owner to be the
croupier has been removed, as it is useless.

* Roulette tweaks and piff

Here's the piff, but it's really hard for me to understand why the
object seems to reset itself.

The only thing I didn't include in the piff is an updated subroutine to
check the machine's balance before allowing it to open (...outside
acceptable limits?) since I can't figure out what the object itself
resets its attributes mid game (only while open with an NPC croupier).

* Roulette fix

Restored functionality for closing the Table, and added the
disconnection of the player (something I originally thought was
automatic after the object event)

* Roulette: oops

:oops:

* Roulette Penultimate Patch

• Machine/Table Balance Textfield and UI Alert Deposit/Withdraw Text
Entry no longer truncate after 5 characters, displaying more than $99999
• Owner Alert now properly displays the minimum balance needed for a
table to open when prompting the owner for their deposit/withdraw amount
• Tables now properly check that the minimum bet is met before accepting
a player's bets. An error is displayed if they player falls short of the
minimum bet
• Roulette plays sounds when a bet is accepted and when a player wins
any money
• "Manage" interaction by owner is now ALWAYS available, which will
prevent tables from becomming stuck and requiring the lot to reset.
Examples of such interruptions include fires and desyncs.
• Multiple tables can be opened and closed and the NPC properly vanishes
upon a table's closing
• A bug was fixed that allowed the owner to set a maximum bet despite
not having enough money in the table to cover it.
• Removed a redundant CloseTable() method call.
• Players cannot play on a table currently being managed by the owner.
Roommates cannot open a table currently being managed by the owner.

* Sound  QoL Update 1: Instant Messages

Just sounds for instant message. Also new message windows start
minimized as they did in TSO.

* FINAL Roulette Patch

•proper piff for machine limits
•machines now limited to $999999 instead of Int32.MaxValue
•as per tsomatt, corrected a sound effect I incorrectly remembered from
10 years ago