Roulette EOD - multiple open tables error
For some reason, MOMI NPCs after the first one do not connect to the
EOD. So a failsafe method of checking the object for a connected NPC was
implemented. This allows any open tables to be playable, rather than
before when the tables assumed they were not open due to a lack of NPC
Avatar.