Hopefully temporary table NSF Alert
Players now know why they were kicked from a table that doesn't have
enough money to cover the maximum bet. Hopefully this is temporary as
ideally a check tree can hide the interaction to open the table, but I
can't get the correct object within the blackjack master object that has
the balance.