Blackjack EOD (#96)
* Blackjack EOD

Introducing Blackjack EOD.

Huge special thanks to Raeven for her expertise on slots and routing.
She also created the dealer object from scratch.

Special thanks to S1ndle for the catalog strings.

Important Note:
While the game (seems to be, hopefully) is 100% working and ready to be
tested in the beta, the dynamic sprites do not yet work. This will be
patched an a future update.

Known Issues:
•Dynamic sprites aren't yet implemented - the table sprites don't
change.
•The dealer won't obey Animate Sim prims to "turn body toward stack
object" so he just faces the table all the time.

<P g="4305B84E" s="15" p="50000" n="Ace Blackjack Table" />

* Remove forced card draws

Forgot to remove some forced card draws in order to test split. Fixed.

* Hopefully temporary table NSF Alert

Players now know why they were kicked from a table that doesn't have
enough money to cover the maximum bet. Hopefully this is temporary as
ideally a check tree can hide the interaction to open the table, but I
can't get the correct object within the blackjack master object that has
the balance.