Add Rotation from any Origin to UIElement
UIImage is the only UIElement to use rotation so far, but it can be used in other classes in the future.

When a rotation angle (radians) is specified that isn't the default 0F, the UImage uses the DrawLocalTexture that allows for a rotation argument. The default rotation Origin is (0.0f, 0.0f) but a custom rotation origin can be set.

Note that the rotation origin specified is not used in the Draw method of the spritebatch, but only to recalculate the 2DMatrix of the UIElement.

Also a change was made (really by ripinperiperi) to HitTestArea.