This change prevents a player from accidently deleting an item when setting the SalePrice via the delete key. Most people are familar with using a combination of delete and backspace in-order to edit text in a textfield, this will prevent the accident that everyone has done at least once, of pressing delete and selling a very expensive item.
Additionally one very small UI/UX enchancement was made, in that when you select "Set for Sale", and the dialog is presented, your cursor is automaticlly focused in and ready to adjust the price.

Implementation: I actually really thought about the best way to take care of this situation. My initial implemntation used on focus events to track the dialog and the focus of the ForSalePrice UITextEdit. But in the end I wanted to make sure that the PriceDialog was getting deallocated correctly, and remove focus. The submitted code seemed like the best of all worlds for memory management and user experience.

Tested in various states: Placing Multiple Items, Setting Multiple Items for Sale, Deleting after an item is set for sale or canceled.