New Jobs UIJobInfo Dialog / Alert (#136)
* Adding some preliminary code that will allow a player to see how many rounds they have left in their current job grade.

* Learning how to make UI elements, tricky, but fun !

* Updating style and starting to finalize the new UIJobInfo panel.

* minor clean up of UIProgressBar.

* Exploring Lot Control, need to refactor some of my code into UIJobInfo

* Minor UI tweaks, phone job info now gives you the same detail as the profile page, minor refactoring of how data is entered into the UI class.

* Minor refactoring.

* Updating the UI to use widely used "inner box" style background on my sections.

* Minor clean up and updating the style of the progress bar.

* FSOFacade: Updating the location of the profile, allowed us to use the previous 'using FSO.Client.UI.Panels' declaration.
UILotControl: Removed hack for grabbing MessageStringID 19. Updated jobcontroller.piff to set the type from message to FSOJOB.
UIPersonPage: Updated the using directive location to be more appropriate.
UIJobInfo: Updated the namespace of location of this class; Corrected spelling mistake of Destroy. Removed default color and shadow sets on yellow text fields; calling Dispose on bespoke texture for progress bar. Remove redeleration of 'ShowDialog' in favor of using the UIScreen.GloablShowDialog method.
VMDialogPrivateStrings: Set this to 127 to prove that I knew what I was modifying, and just left it here. It IS set to 127 in the included piff change.