Fixing Accidental Sale from Price Dialog (#137)
* This change prevents a player from accidently deleting an item when setting the SalePrice via the delete key.
Most people are familar with using a combination of delete and backspace in-order to edit text in a textfield, this will prevent the accident that everyone has done at least once, of pressing delete and selling a very expensive item.

Additionally one very small UI/UX enchancement was made, in that when you select "Set for Sale", and the dialog is presented, your cursor is automaticlly focused in and ready to adjust the price.

Implementation: I actually really thought about the best way to take care of this situation. My initial implemntation used on focus events to track the dialog and the focus of the ForSalePrice UITextEdit. But in the end I wanted to make sure that the PriceDialog was getting deallocated correctly, and remove focus. The submitted code seemed like the best of all worlds for memory management and user experience.

Tested in various states: Placing Multiple Items, Setting Multiple Items for Sale, Deleting after an item is set for sale or canceled.

* Returning the dialogs scope. Removed extra delegation on UIAsyncPriceDialog. Checking if focus is null. There is a slight caveat to this. In that you have to not be holding the object (you have to let it go), in order for the focus to clear. There is a cost/benefit to this however, in that this solution now prevents delete in ANY text input from accidently deleting the object. But again, means you have to put the object down. To help faciliate this, after you mark an object forsale, i'm now canceling the pick up action (you already had to put it down once in order for this to even be marked forsale). So the extra benefit is now it is slightly easier (one less click) to pricing a lot of object quickly.