[Server] More lot trading/move improvements
- Fix placing a Mayor lot resetting move date. (therefore making it
impossible to post bulletins or run for the next election)
- Trading a property now correctly updates move date, only when the user
is trading themselves into a new neighborhood.
- Trading a property can no longer include untradable objects. (they are
sent back to the owner)
- (note, all lot move modes have been verified to only reset your move
date when you are moving to a new neighborhood, or from not having any
property. the best way to avoid issues is to switch lots with secure
trade, or move via city view)