[Server] Various fixes.
- Only send the alert that all roomies have left the property when their
sims have not been deleted yet. (avoids dialog duplication and showing
the dialog to the roomie on rejoin)
- Timeout connections that have not provided authentication (server,
stops leaked connections)
- Fix a bug that would force you out of the property on showing the
timeout dialog (if your window is too small)
- Fixed attempting to purchase too many clothes. Now refunds the
purchaser if the outfit owner cannot be changed (not the best solution,
no error msg but works for now)
- Wait 100ms before re-establishing connection, to deal with cases where
we are yet to receive the ServerByePDU or are disconnecting ourselves.
- Draw lot facades after the flashing square beneath them (when they are
online) to avoid masking issues.
- Don't cut off city object opacity at 0.1.
- City tree alpha is now non-premultiplied.
- Fix object alpha blending for surrounding lots (roughly)