[Client] One connection at a time
AriesClients can now only have one active connection. Attempting to
connect a second time will immediately close the first connection.
(stops the game from monitoring both an attempted reconnect to a
previous lot, and a connection to the current lot at the same time. This
was causing the game to believe it lost connection to the current lot
when the attempted reconnect timed out and disconnected.)