CooldownEOD (#161)
* HoldemCasinoEOD plus a few updates to older EODs and some misc

Full description in pull request to main repo, relevant strings for
global server listed elsewhere

* HoldemCasino EOD: Logic Error

Deal hole cards face up except dealer. Duh

* Console Removal

Forgot to remove Console.WriteLine

* S1ndle's Amazing Casino Set

<P g="02B9B1A7" s="19" p="305" n="lamp-ceiling-Hold'em Dice lamp" />

<P g="02B9B1A8" s="19" p="265" n="lamp-floor-Hold'em Dice lamp" />

<P g="02B9B1A9" s="19" p="175" n="lamp-table-Hold'em Dice lamp" />

<P g="02B9B1AA" s="19" p="230" n="lamp-wall-Hold'em Neon Sign" />

<P g="02B9B1AC" s="13" p="350" n="counter-1-The Hold'em Cabinets" />

<P g="02B9B1AD" s="13" p="325" n="counter-2-The Hold'em Island Counter"
/>

<P g="22EF511F" s="13" p="1000" n="table-end-The Hold'em Display Table"
/>

<P g="31CF873D" s="17" p="485" n="rug-runner-1x4-The Hold'em Runner" />

<P g="0D79D716" s="17" p="700" n="rug-area-3x3-The Hold'em Area Rug" />

<P g="7E055DC7" s="18" p="115" n="sign-foldable-The Hold'em Sign" />

<P g="20B11DFB" s="7" p="15" n="column-The Hold'em Column" />

<P g="7CEF5108" s="18" p="1100" n="bar-The Hold'em Watering Hole" />

The Game Table itself:
<P g="37EF5198" s="15" p="35000" n="Casino Hold'em" />

* #113 Review

•Required fix completed
•Nearly all suggested additional changes completed
•Found another logic error related to face up cards
•Found a nasty bug that would have caused players to push each other off
of the table, and they probably were doing so on the blackjack table, so
fixed it there, too

* Final? changes

• Nuget HandEvaluator
• Requested locks added
• Couple of null checks added, though not necessary since correct nuget
package was added

* Jams

:jams:

* HoldemCasinoEOD & BlackjackEOD Patch

• Client: Null check for very rare bug
• Strings changed from "Play Blackjack" to "Play" for both tables
• Check for NPC Dealer before trying to push interaction and possibly
throwing null

* TwoPersonJobObjectMazeEOD

Vanilla TSO Maze, with a twist, plus:
• UIImages can have tooltips

* Of course you know this means war...

Creative, right?

* Unleashed Coffee Table fix

Fix for unleashed coffee table (inits for left and right tiles are now
properly assigned and labelled).

Last commit I promise.

* CT - Is Skill Object? fix

sorry

* Bugfixes: BlackjackEOD & HoldemCasinoEOD

•The repairman bug (infinite loop) is now fixed for HoldemCasino tables
•Pushing a hand on which a player doubled down now returns the doubled
bet amount, rather than just the original bet amount

* BlackjackEOD: Min Table Balance reworked

Needs to be 8 times maximum bet instead of 6 to cover each player
splitting every hand and getting a blackjack on all of them.

•So worst case scenario is 4 players with 4 blackjacks, less 4 times
their initial bet: 1 for the original bet and 3 for each split.
•MinimumBet * (3:2 BlackjackWinRatio) * 4(Hands) * 4(Players) -
MinimumBet * 4 (Hands) * 4(Players) = 8*MinimumBet = TableMinBalance

* Obstensible Fixed Side Bet

Changed side bet payout logic to make sure only one player is paid for
their side bet, based on bug reports of "phamtom payouts"

The bug reports were not sufficient in helping me uncover the error, but
these fixes are good failsafes.

* No more double dealing - Blackjack, HoldemCasino, Roulette

Attempt 1 to fix the double NPCs phenomenon: a remove object instance
prim is invoked at the end of the Dealing/Croupier: Stop interaction(s).

* New EOD Panel Size

Future EODs of DoubleTall size won't have to be bound by the seam..

* New Seasonal Costume

A seasonal costume is being added to the black "costume" trunk for both
genders.

* Null Check for OnClose

A bug was encoutnered in BandEOD where a Sim would disconnect from the
EOD server but the EOD Panel would not close. A patch for band is
forthcoming, but closing the EODPanel shouldn't throw null, either.

* piffsfromdiscord

* missedone

* bugfix

* Add Rotation from any Origin to UIElement

UIImage is the only UIElement to use rotation so far, but it can be used in other classes in the future.

When a rotation angle (radians) is specified that isn't the default 0F, the UImage uses the DrawLocalTexture that allows for a rotation argument. The default rotation Origin is (0.0f, 0.0f) but a custom rotation origin can be set.

Note that the rotation origin specified is not used in the Draw method of the spritebatch, but only to recalculate the 2DMatrix of the UIElement.

Also a change was made (really by ripinperiperi) to HitTestArea.

* UI Updates

Boring updates to the UI, but maybe helpful some day

* Update UISlotsEOD.cs

Wheelspin animations are faster to match the 1.2x animation speed from ages ago.

* Update UIInvisibleButton.cs

Feedback from cbt, less expensive version of invisible button

* Make Band Suck Less - Updates to band

Pending cbt test

* UIImage - Remove DissolveOpacity

Remove DissolveOpacity for PR

* Band piff & Tex

.piff and Textures for Band 2

* Update CityGeometry.cs

City geometry fix - won't error on non-Sunrise crater maps

* Band strings

Band2 strings should be in their own file.

* Revert City Geometry

Reverted to current deployed version as requested

* Band update from CBT Feedback

•Payout has been boosted, especially at early rounds
•RockOn or Sellout? ends when 3 people choose sellout or two people choose rock on.
•Failsafes to keep buttons from appearing activated
•Failsafes added to cleanly go back to the lobby when a player disconnects
•Note tooltips are no longer solfeggio, but are now E major notes with a Lo and Hi E (hard coded literals)

* uh ok

* band strings mismatch

hate csproj files

* Cooldown EOD

Initial commit. No object with which to test yet, so entries for EOD don't exist yet in VMEODServer and UIEODController. So this isn't actually able to be tested yet.

* CooldownEOD final fixes

Ready for testing

* 0x00001004

oh yea