[VM, LotView] Lots of bugfixes
- Interaction Priority is now tracked similar to TS1. Priority within
the queue is not mutable - it is copied to the Person Data "priority"
when the interaction is entered, and that is treated as the current
priority by the queue. Leaving an interaction's context will set the
priority to the new interaction's priority, or 0 if there are none left.
- Greatly reworked FindBestAction. Still work in progress - need curves
for interaction scores. (see globals.iff)
- Fixed an issue where the ground would offset up or down one pixel
randomly in 2D mode.
- Interactions with an invalid check tree are now ignored.
- Objects with functional score <= 0 are now ignored when considering
them. (eg. for wash dishes function) Fixes a few objects with invalid
OBJf.
- Add "magic" headline type for TS1
- Changed lots of TS1 behaviour. Still WIP.
- Fix a really specific issue where if an object deleted itself on the
first frame the lot opened, it would break the scheduler.
- Some more changes

TSO is likely broken due to this.