[VM] Entirely reworked autonomy.
- Autonomy has been reversed from values provided from The Sims 1.0.
This will affect TS1 and TSO, and finally autonomous sims are behaving
somewhat reasonably. Optimisations to come.
- TS1Curves for autonomy are read from global.iff. For these to be
present, you cannot be on the initial version of the game. May have
hardcoded variant in future.
- See VMFindBestAction.cs for more details.
- Volcanic non longer creates multiple instances each time you switch
lots. It will use the same window (and open objects), but close any
tracers that were open in the previous lot.
- Volcanic now closes with the game.
- Added "sleep state -1" as used by pets in TS1.
- Fixed an issue where Cat and Dog interactions wouldn't work in TS1
because their flags overlapped with special TSO ones.
- Do not center to sims offscreen.
- Fixed generating TS1 thumbnails in 2D.
- Burn primitive can no longer burn out of world.
- Fixed TS1 objects selling for 0 or the wrong value entirely
(deprecation still not impl)
- Sims now wake up in TS1 on the wake hour if their motive maxes (7am)
- Fixed a bug where interaction priority would not be reset when a sim
resets.
- Improved TS1 catalog logic.
- Fixed 3D thumbnail jumbling multitile parts if there are empty
segments.
- STR editor in Volcanic now shows comments.
- FCNS chunk is now parsed (ver 1 + 2)