Complete OBJM person struct, HouseSpy
House Spy allows you to open a house iff and see person state contained within. You can play and save the house in TS1 (press ctrl-s), and the display on House Spy will instantly update. The current purpose is for seeing how the unknown values change based on interactions ingame, but it could be expanded to do more.